+++ GrowDay.Domain/DTO/CreateSuggestedHabitDTO.cs
@@ -13,8 +13,8 @@ namespace GrowDay.Domain.DTO
         public DateTime? StartDate { get; set; }
         public DateTime? EndDate { get; set; }
 
-        public int? TargetValue { get; set; }
-        public int? IncrementValue { get; set; }
+        public double? TargetValue { get; set; }
+        public double? IncrementValue { get; set; }
         public string? Unit { get; set; }
     }
 }
