+++ GrowDay.Domain/Entities/Concretes/Habit.cs
@@ -17,8 +17,8 @@ namespace GrowDay.Domain.Entities.Concretes
         public bool IsGlobal { get; set; } = false;
         public string? CreatedByUserId { get; set; }
 
-        public int? TargetValue { get; set; }
-        public int? IncrementValue { get; set; }
+        public double? TargetValue { get; set; }
+        public double? IncrementValue { get; set; }
         public string? Unit { get; set; }
 
         //Navigation Property
