+++ GrowDay.Domain/Entities/Concretes/UserHabit.cs
@@ -20,9 +20,9 @@ namespace GrowDay.Domain.Entities.Concretes
         public int LongestStreak { get; set; }
         public DateTime? LastCompletedDate { get; set; }
 
-        public int? CurrentValue { get; set; }
-        public int? TargetValue { get; set; }
-        public int? IncrementValue { get; set; }
+        public double? CurrentValue { get; set; }
+        public double? TargetValue { get; set; }
+        public double? IncrementValue { get; set; }
         public string? Unit { get; set; }
 
         public TimeSpan? NotificationTime { get; set; }
